Precision overhaul and bug fixes #97

+++ src/com/aryan/rain/Game.java
@@ -139,10 +139,10 @@ public class Game extends Canvas implements Runnable{
         }
         screen.clear();
 
-        int xScroll = player.getX() - screen.width / 2;
-        int yScroll = player.getY() - screen.height / 2;
+        double xScroll = player.getX() - screen.width / 2;
+        double yScroll = player.getY() - screen.height / 2;
 
-        level.render(xScroll, yScroll, screen);
+        level.render((int)xScroll, (int)yScroll, screen);
         // player.render(screen);
 
         //////////////////////// Particle graphics demo ////////////////////////
