+++ src/com/aryan/rain/level/Level.java
@@ -195,13 +195,13 @@ public class Level {
     public List<Entity> getEntities(Entity e, int radius){
         List<Entity> result = new ArrayList<>();
 
-        int ex = e.getX();
-        int ey = e.getY();
+        int ex = (int)e.getX();
+        int ey = (int)e.getY();
 
         for (int i=0; i<entities.size(); i++){
             Entity entity = entities.get(i);
-            int x = entity.getX();
-            int y = entity.getY();
+            int x = (int) entity.getX();
+            int y = (int) entity.getY();
 
             int dx = Math.abs(x - ex);
             int dy = Math.abs(y - ey);
@@ -220,13 +220,13 @@ public class Level {
 
 //        List<Entity> entities = getEntities(e, radius);
         List<Player> result = new ArrayList<Player>();
-        int ex = e.getX();
-        int ey = e.getY();
+        int ex = (int) e.getX();
+        int ey = (int) e.getY();
 
         for (int i=0; i<players.size(); i++){
             Player player = players.get(i);
-            int x = player.getX();
-            int y = player.getY();
+            int x = (int) player.getX();
+            int y = (int) player.getY();
 
             int dx = Math.abs(x - ex);
             int dy = Math.abs(y - ey);
